// src/app/models/email.ts
export interface Email {
    _id: string;
    email_name: any;
    email_text: {
        paragraph1: string;
        paragraph2: string;
        paragraph3: string;
        paragraph4: string;
        paragraph5: string;
        paragraph6: string;
        paragraph7: string;
        paragraph8: string;
        paragraph9: string;
        paragraph10: string;
    };
    created_by: any;
    modified_by: any;
    modified_date: any;
}